refactored ui into ui folder, refactored inputBox into its own file
