+++ index.html
@@ -7,7 +7,7 @@
     <title>Pupupeli</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="css/style.css">
-    <script type="module" src="src/ui.js"></script>
+    <script type="module" src="src/ui/ui.js"></script>
 </head>
 
 <body>
